Move modules related to validation into own subpackage.
