+++ pyaml/accelerator.py
@@ -8,10 +8,8 @@ from .arrays.array import Array, ArraySchema
 from .common.element import Element, ElementSchema
 from .common.element_holder import ElementHolder
 from .common.exception import PyAMLConfigException
-from .configuration import ConfigurationManager, UnsupportedConfigurationRootError
-from .configuration.configuration_models import ConfigurationSchema
+from .configuration import ConfigurationManager, ConfigurationSchema, UnsupportedConfigurationRootError, register_schema
 from .configuration.factory import Factory
-from .configuration.schema_registry import register_schema
 from .control.controlsystem import ControlSystem, ControlSystemSchema
 from .lattice.simulator import Simulator, SimulatorSchema
 from .yellow_pages import YellowPages
