+++ pyaml/configuration/__init__.py
@@ -1,19 +1,17 @@
 """
-PyAML configuration module
+PyAML configuration subpackage.
 """
 
-from .configuration_models import ConfigurationSchema
 from .factory import Factory
 from .fileloader import get_root_folder, set_root_folder
 from .manager import ConfigurationManager, UnsupportedConfigurationRootError
-from .schema_registry import SchemaRegistry, register_schema
+from .validation import ConfigurationSchema, register_schema
 
 __all__ = [
     "ConfigurationManager",
     "ConfigurationSchema",
     "UnsupportedConfigurationRootError",
     "Factory",
-    "SchemaRegistry",
     "get_root_folder",
     "set_root_folder",
     "register_schema",
