+++ pyaml/configuration/validation/legacy_registry.json
