+++ pyaml/control/controlsystem.py
@@ -10,7 +10,7 @@ from ..common.abstract_aggregator import ScalarAggregator
 from ..common.element import Element
 from ..common.element_holder import ElementHolder
 from ..common.exception import PyAMLException
-from ..configuration.configuration_models import ConfigurationSchema
+from ..configuration import ConfigurationSchema
 from ..configuration.factory import Factory
 from ..configuration.unbound_element import UnboundElement
 from ..control.abstract_impl import (
