+++ pyaml/lattice/simulator.py
@@ -8,9 +8,7 @@ from ..common.abstract_aggregator import ScalarAggregator
 from ..common.element import Element
 from ..common.element_holder import ElementHolder
 from ..common.exception import PyAMLException
-from ..configuration import get_root_folder
-from ..configuration.configuration_models import ConfigurationSchema
-from ..configuration.schema_registry import register_schema
+from ..configuration import ConfigurationSchema, get_root_folder, register_schema
 from ..diagnostics.tune_monitor import BetatronTuneMonitor
 from ..lattice.abstract_impl import (
     BPMHScalarAggregator,
