+++ pyaml/magnet/identity_model.py
@@ -3,8 +3,7 @@ from pydantic import BaseModel, ConfigDict
 
 from .. import PyAMLException
 from ..common.element import __pyaml_repr__
-from ..configuration.configuration_models import ConfigurationSchema
-from ..configuration.schema_registry import register_schema
+from ..configuration import ConfigurationSchema, register_schema
 from ..control.deviceaccess import DeviceAccess, DeviceAccessSchema
 from .model import MagnetModel
