+++ pyaml/magnet/linear_serialized_model.py
@@ -3,11 +3,10 @@ from pydantic import BaseModel, ConfigDict
 
 from ..common.element import __pyaml_repr__
 from ..common.exception import PyAMLException
-from ..configuration.configuration_models import ConfigurationSchema
+from ..configuration import ConfigurationSchema, register_schema
 from ..configuration.curve import Curve, CurveSchema
 from ..configuration.inline_curve import InlineCurve, InlineCurveSchema
 from ..configuration.matrix import Matrix
-from ..configuration.schema_registry import register_schema
 from ..control.deviceaccess import DeviceAccess, DeviceAccessSchema
 from .linear_model import LinearMagnetModel, LinearMagnetModelSchema
 from .model import MagnetModel
