+++ pyaml/magnet/serialized_magnet.py
@@ -4,8 +4,7 @@ from scipy.constants import speed_of_light
 from .. import PyAMLException
 from ..common import abstract
 from ..common.element import Element, ElementSchema, __pyaml_repr__
-from ..configuration import Factory
-from ..configuration.schema_registry import register_schema
+from ..configuration import ConfigurationSchema, Factory, register_schema
 from ..control.deviceaccess import DeviceAccess
 from .function_mapping import function_map
 from .magnet import Magnet
