+++ pyaml/magnet/spline_model.py
@@ -3,9 +3,8 @@ from pydantic import BaseModel, ConfigDict
 from scipy.interpolate import make_smoothing_spline
 
 from ..common.element import __pyaml_repr__
-from ..configuration.configuration_models import ConfigurationSchema
+from ..configuration import ConfigurationSchema, register_schema
 from ..configuration.curve import Curve, CurveSchema
-from ..configuration.schema_registry import register_schema
 from ..control.deviceaccess import DeviceAccess, DeviceAccessSchema
 from .model import MagnetModel
