+++ pyaml/magnet/vcorrector.py
@@ -1,5 +1,5 @@
 from ..common import abstract
-from ..configuration.schema_registry import register_schema
+from ..configuration import register_schema
 from ..lattice.polynom_info import PolynomInfo
 from .corrector import RWCorrectorAngle
 from .magnet import Magnet, MagnetSchema
