+++ pyaml/rf/rf_plant.py
@@ -9,7 +9,7 @@ except ImportError:
 from .. import PyAMLException
 from ..common import abstract
 from ..common.element import Element, ElementSchema
-from ..configuration.schema_registry import register_schema
+from ..configuration import register_schema
 from ..control.deviceaccess import DeviceAccess, DeviceAccessSchema
 from .rf_transmitter import RFTransmitter, RFTransmitterSchema
