+++ pyaml/tuning_tools/response_matrix_data.py
@@ -7,9 +7,8 @@ from pyaml.common.element import __pyaml_repr__
 from pyaml.configuration.factory import Factory
 
 from .. import PyAMLException
-from ..configuration.configuration_models import ConfigurationSchema
+from ..configuration import ConfigurationSchema, register_schema
 from ..configuration.fileloader import load
-from ..configuration.schema_registry import register_schema
 
 
 class ResponseMatrixDataSchema(ConfigurationSchema):
